requirements.txt
opencv-python
numpy
scipy
pyyaml
ultralytics
torch
torchvision
transformers
segment-anything
